COMMENT OUT RESETS IN GRIDDY.CSS
CSS order is being rearranged, so that drop-in.css loads BEFORE
griddy.css. For this reason, the resets (link to normalize, code
from Paul Irish and HTML5 Boilerplate, etc.) must be moved to
drop-in.css. This merge comments out these resets in griddy.css
for testing without deleting them entirely.